+++ arithmetic_handlers.c
@@ -1,0 +1,54 @@
+#include "monty.h"
+
+
+/**
+ * handle_add - adds the top two elements and stores
+ * the results in the first node
+ * @stack: a pointer to the stack
+ * @line_number: position of the opcode in the file
+ */
+void handle_add(stack_t **stack, unsigned int line_number)
+{
+
+	int result;
+	stack_t *first, *second;
+
+	if (stack_len(*stack) < 2)
+	{
+		fprintf(stderr, "L%d: can't add, stack too short\n", line_number);
+		free_stack(*stack);
+		exit(EXIT_FAILURE);
+	}
+
+	first = node_at_index(*stack, 0);
+	second = node_at_index(*stack, 1);
+
+	result = first->n + second->n;
+
+	second->n = result;
+	(*stack) = second;
+
+	free(first);
+}
+
+
+void handle_sub(stack_t **stack, unsigned int line_number)
+{
+	(void)stack;
+	(void)line_number;
+}
+void handle_div(stack_t **stack, unsigned int line_number)
+{
+	(void)stack;
+	(void)line_number;
+}
+void handle_mul(stack_t **stack, unsigned int line_number)
+{
+	(void)stack;
+	(void)line_number;
+}
+void handle_mod(stack_t **stack, unsigned int line_number)
+{
+	(void)stack;
+	(void)line_number;
+}
