+++ utilities.c
@@ -1,5 +1,19 @@
 #include "monty.h"
 
+/**
+ * set_arg - retrieve opcode argument from the environment variables
+ * @arg: the variable on which to cast the argument
+ * @command: the command being executed
+ */
+void set_arg(char *arg, char *command)
+{
+	int index = strlen(command);
+
+	while (strchr(&monty_opcode[index], ' ') == 0)
+		index++;
+	strncpy(arg, monty_opcode + index, 1024);
+	strip(arg, NULL);
+}
 
 /**
  * strip - remove specific characters at the beginning and end of string
